refactor(format): extract cyclonedx export owner
