+++ ci/proof.toml
@@ -723,6 +723,7 @@ kind = "rust"
 paths = [
   "crates/tokmd-format/src/diff/**",
   "crates/tokmd-format/src/diff.rs",
+  "crates/tokmd-format/src/export/**",
   "crates/tokmd-format/src/export.rs",
   "crates/tokmd-format/src/lib.rs",
   "crates/tokmd-format/src/summary.rs",
